[Framework] Allow status property in spec description (#269)
See #6.

The `status` property can now be added to the description of a spec to override
or specify the maturity level of the spec.

The list of values is currently constrained to W3C maturity levels. We may want
to add additional maturity levels later on, e.g. those used at IETF.