[Framework] Support mdn-browser-compat-data as source
See https://github.com/w3c/web-roadmaps/issues/197

With this update, the framework supports a new `mdn` source that tells it to
fetch implementation data from the MDN Browser Compatibility Data.

Under the hoods, the code uses the `mdn-browser-compat-data` package.

Note that, given that the granularity of the implementation data in the MDN
browser compat project is finer than the spec level, it is likely that, for
most specs, we'll need to define sub-features that link to individual MDN data
files (or to sub-features within these data files).

README updated to explain how to reference MDN browser compatibility data and
how to define features in a spec.