[About page] Add partial badge description
The badge gets displayed when the implementation info is not complete, typically
because we have implementation info for some features in the spec but not for
the entire spec. The badge conveys the fact that support may be partial as a
result.

NB: text in the Chinese page needs to be translated.