Framework specfeature (#270)
* [Framework] Support linking to more than one ID on status platform

See https://github.com/w3c/web-roadmaps/issues/59

The description of a specification can now list features that compose the spec
along with info about where to extract implementation status for each feature
on status platforms.

For instance, to say that the CSS Backgrounds spec contains the features
"Rounded corners" and "Backgrounds", one can define the spec as:

```json
{
  "url": "https://www.w3.org/TR/css-backgrounds-3/",
  "impl": { ... },
  "features": {
    "border-radius": {
      "url": "#the-border-radius",
      "title": "Rounded corners",
      "impl": { ... }
    },
    "background": {
      "url": "#the-background",
      "title": "Backgrounds",
      "impl": { ... }
    }
  },
  "featuresCoverage": "partial"
}
```

The `featuresCoverage` flag is set to `partial` by default. If set to `full`,
and if we have implementation info from one source on all features for a given
browser, then the framework will guess that the browser supports the entire
spec.

This new mechanism replaces the previous `feature` property (that previous
mechanism can still be used, although it is deprecated).

When support may be partial, an orange banner gets displayed on top of the
implementation icon to convey the possibility that the implementation status
may not apply to the whole spec.

* [Framework] Support mdn-browser-compat-data as source

See https://github.com/w3c/web-roadmaps/issues/197

With this update, the framework supports a new `mdn` source that tells it to
fetch implementation data from the MDN Browser Compatibility Data.

Under the hoods, the code uses the `mdn-browser-compat-data` package.

Note that, given that the granularity of the implementation data in the MDN
browser compat project is finer than the spec level, it is likely that, for
most specs, we'll need to define sub-features that link to individual MDN data
files (or to sub-features within these data files).

README updated to explain how to reference MDN browser compatibility data and
how to define features in a spec.

* [About page] Add partial badge description

The badge gets displayed when the implementation info is not complete, typically
because we have implementation info for some features in the spec but not for
the entire spec. The badge conveys the fact that support may be partial as a
result.

NB: text in the Chinese page needs to be translated.