[Framework] Add mechanism to link to the ED
See https://github.com/w3c/web-roadmaps/issues/106

3 new mechanisms to link to the Editor's Draft:
1. add a `data-linkto="ED"` attribute next to the `data-featureid` attribute to
have the framework link to the Editor's Draf for that link
2. add a `linkto` property set to `ED` to the `toc.json` file to have the
framework link to the Editor's Draft for all links in the prose.
3. define custom tables and set the `linkto` property of the `spec` column to
`ED` to tell the framework to link to the Editor's Draft in the summary table.

Note the update also fixes fragment handling when features are targeted (the
URL of the feature was previously unused, meaning links always targeted the
top of the spec, and not the relevant section when a feature was used).