[data] impl updates and features creation
See #271.

Major pass through all the data files to:
- complete the references to the different status platforms;
- add references to MDN Browser Compatibility data in particular;
- create subfeatures when needed;
- move the URL to the first line;
- use `url` consistently throughout the data files

The update is certainly not perfect. I did not try to merge all HTML5 features
into one data file for instance, although that could be done. There remain a
couple of cases where the level for which we have the implementation
data is not necessarily the level that we're describing. Also, in some cases,
I took the liberty to consider that referencing the implementation data for the
main interface that a spec on MDN is enough. All in all, the info should be
more complete than before though.

Note the update also fixes a couple of bugs in `extract-impl-data.js`.