[Mobile] Mention overscroll-behavior
The `overscroll-behavior` property is useful e.g. to implement pull-to-refresh
and infinite scrolling efficiently. Mentioned in Performance and tuning.

Note the draft is still undergoing transition from the WICG to the CSS WG. The
URL of the spec may need to be updated when that transition is complete (and
the mention may need to be moved out of the "Exploratory work" section).

Fix #258.