[Framework] Fix URL generation for Specref reverse lookup
The lookup URL that was created dropped the final HTML filename for some reason
for non-TR specs, but Specref needs it to identify the spec that the request
is about.

For instance, Specref returns useful info for:
`https://wicg.github.io/page-lifecycle/spec.html`

... but won't return anything for:
`https://wicg.github.io/page-lifecycle/`

The final HTML filename is now kept.