Generate group info automatically from publisher when possible
The spec info generation tool did not properly populate the `deliveredBy`
property from the publisher info. It either failed to do so, or actually
populated the property with the publisher's info even though the publisher was
not a group (e.g. OGC is not a group, it's a consortium made of groups).

See https://github.com/w3c/web-roadmaps/pull/342#pullrequestreview-183114990