Generate group info automatically from publisher when possible (#343)
The spec info generation tool did not properly populate the `deliveredBy`
property from the publisher info. It either failed to do so, or actually
populated the property with the publisher's info even though the publisher was
not a group (e.g. OGC is not a group, it's a consortium made of groups).