[Mobile] Fix Geolocation Sensor text in Application Lifecycle (#355)
In https://github.com/w3c/web-roadmaps/pull/341#issuecomment-445884755
I suggested to update "Application Lifecycle". The update was done... to the
games roadmap, not to the mobile roadmap! This update brings the changes to
the right page.

Updating the text in the games roadmap is good too. I'll push other updates to
the games roadmap later on and separately.