[mobile] Mention Resize Observer (#368)
Resize Observer is a solution for what was called "Element queries" in the
roadmap. The solution is not declarative as initially envisioned (i.e. not a
media query), but it allows the creation of responsive Web components. The text
makes the link between the two explicit and removes the entry for "Element
queries" (the text still points at the use cases document).

Also, the text now appears in the "In progress" section, given that the spec has
been adopted by the CSS WG (no FPWD yet, though).