Compute proper Specref lookup ID for multipage specs
We're going to have a number of data files that reference specific features in
the HTML LS spec. To avoid having to repeat the spec's title in each data file,
the function that determines the lookup ID to use with Specref now drops the
final page at the end of the URL to end up with
`http://html.spec.whatwg.org/multipage` which Specref understands.