[mobile] Properly mention WebGL and WebGPU in Graphics (#438)
Mentions of WebGL and WebGPU were in a note and did not trigger the creation of
entries in the implementation table. The text in Well-deployed technologies now
properly references the WebGL spec and makes it clear that the implementation
info is for version 1.0. It also links to the Exploratory work section that
now links to the WebGPU draft.

Note I didn't add a second entry point for WebGL version 2.0 for now, that does
not seem necessary.