[media] Refactor app dev into device interoperability
The Media Application development section adopted a developer perspective, but
the topic is also highly relevant for device manufacturers and integrators.

This commit updates the section's title and description to set the context, and
lists additional specifications that developers can use to detect capabilities
of the underlying platform and measure the performance of their applications.