Add createFeatureSections param and enable it in mobile roadmap
Paragraphs and inner section may talk about a given feature. That feature gets
displayed in the summary table at the end of the section but didn't appear
inline.

Given the way we build roadmaps, there is very little continuity in the prose
when we switch from one feature to another. As sections are becoming longer,
this makes negatively affects readability.

The update adds a `createFeatureSections` parameter that can be set in the
`toc.json` file to enable the automatic creation of sections and headers for
divs and paragraphs that have a `data-feature` attribute. This allows to group
content together visually under a feature heading.

On small displays, this feature heading is displayed inline (under the hoods,
it's an `h3` element). On large displays, it appears on the left to make the
page less dull.

I adjusted several pages of the mobile roadmap to make things consistent.

README updated as well to mention the setting.