[media] Enable features sections (#452)
Sets the new `createFeatureSections` flag to have the framework generate
feature sections to visually group paragraphs that talk about a same overall
feature under a section heading. See:
https://github.com/w3c/web-roadmaps#creating-feature-sections

Except for the first paragraph in the "Rendering" page, all pages that compose
the roadmap already had `data-feature` attributes in `<div>` or `<p>` elements,
so no need to update the actual contents to take advantage of the flag.