Add groups view (#458)
The view gets generated from all features referenced throughout the pages. It
needs to be enabled in the toc.json file. This view matches #127. The Media &
Entertainment Interest Group recently requested it again to create some sort of
"heat map" of what's going on.

The list is pretty raw for now: it features the name of the group, the relevant
specifications (or features in these specifications) that the group is working
on (or was working on), and links to pages where the specifications are
mentioned for further details.

Changes to data files were needed to align group names with their official
names.

README not yet updated.