Consolidate groups where needed
Some published documents that are in scope of an active working group are still
associated with a group that no longer exists, because the working group has not
yet published an update to /TR/. That is factually correct, but confusing
especially with the new list of groups view.

This update consolidates groups info to put:
- MSE/EME under the Media WG
- the Open Screen Protocol under the Second Screen WG
- the Canvas API under the WHATWG
- UI Events under the WebApps WG

The update also fixes wrapping issues for lengthy names, typically to prevent
HTMLVideoElement.requestVideoFrameCallback() from overflowing.