Lookup spec in browser-specs, first pass at cleaning data
The code already interfaced with the W3C API and specref but, for non /TR/
specs, this required tracking the URL and sometimes the title in the data file,
which is error-prone (both may evolve as spec progresses).

browser-specs is an up-to-date and curated list of specifications implemented
(or being worked on) by Web browsers. The code now uses at its first source of
information. browser-specs does not contain all relevant specs for roadmaps, but
that's still a better start.

This update includes a first pass at data files to drop now useless "url" and
"title" fields when possible (meaning when the filename matches the shortname
of the spec in browser-specs). A second pass will be needed to update the names
of the files to the spec shortname where possible, as well as to reference
features from within the main spec file instead of maintaining multiple specs
for each of them.

The goal with these updates is to reduce the need to maintain data to a bare
minimum.