Support the new Microsoft Edge in implementation status
Implementation status for Microsoft Edge was mostly still targeting the old
version, not the one based on Chromium, see #447.

This update makes sure that implementation status info for Microsoft Edge is
for the version based on Chromium, and updates the Edge logo accordingly.

In practice, the code can no longer rely on the JSON file that ship with the
Edge Platform Status site, since it only contains information about the previous
version. Information about Microsoft Edge in Chrome Platform Status (which is
probably mostly based on that JSON file as well) also seems to target the
previous version of Microsoft Edge.

The information in "Can I use" and MDN Browser Compat Data seem to be about the
new version and can be used directly.

But what seems to work well is to actually rely on the implementation
information for Chrome from Chrome Platform Status, and to simply check the
version of Chromium that ships with the latest version of Microsoft Edge (this
version is typically off by one compared to Chrome) to assess whether a feature
is available in Microsoft Edge or not.

This is what this update does. It also disables support for Microsoft Edge
Platform Status, at least temporarily (I kept the code to ease rollback later
on).

This update also includes a couple of manual updates to implementation data,
e.g. triggered by outdated data in Chrome Platform Status.